Continue implementing SQLite
On bot startup, creates a table for the User Emojis cog if it doesn't exist already.
When the bot joins/leaves a guild, the guild ID is added/removed as a row in a table.
When a member joins/leaves a guild, the member ID is added/removed as a row in a table.